requirements.txt
asgiref==3.7.2
certifi==2023.7.22
charset-normalizer==3.2.0
Django==4.2.5
django-cors-headers==4.2.0
django-filter==23.2
django-qiniu-storage==2.3.1
django-rest-multiple-models==2.1.3
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.0
idna==3.4
PyJWT==2.8.0
pytz==2023.3.post1
qiniu==7.11.1
requests==2.31.0
six==1.16.0
sqlparse==0.4.4
tzdata==2023.3
urllib3==2.0.4
